Physics Units Cleanup
Cleaning up the current Physics units to match the updated interfaces for Generic math. Backed up the old types in case there is something I missed.
The Excel spreadsheets are still out of date.
Trying out an idea for automatic conversion of units on construction using interfaces.
Adding some null checking because I tried out !! but it was removed in the latest preview.